requirements.txt
firebase-admin
functions-framework
google-cloud-texttospeech
replicate
requests
